refactor: club 멤버 그룹핑/정렬 함수를 파일별로 분리
groupMembersByGrade, getSortedGrades를 각각 독립 파일로 분리해
단일 책임 원칙을 적용하고 Vitest 테스트 대상 명확화
