+++ src/entities/club/lib/getSortedGrades.ts
@@ -1,0 +1,7 @@
+export function getSortedGrades(
+  grouped: Record<number, unknown[]>,
+): number[] {
+  return Object.keys(grouped)
+    .map(Number)
+    .sort((a, b) => b - a);
+}
